+++ src/media/font/bitmap/mod.rs
@@ -1,0 +1,12 @@
+// devela::media::bitmap
+//
+//! Bitmap fonts
+//
+
+#[cfg(test)]
+mod tests;
+
+mod bitmap;
+mod fonts;
+
+pub use {bitmap::*, fonts::*};